Fix GS backface culling
The previous commit wasn't quite right, and the code didn't make
sense either.  It was just getting lucky most of the time and
the faces were getting drawn at the wrong times with closer
examination.

Change pipe to transform points into eye space in the VS, which
means apply just the model and view transform.
Convert the eye location to eye space in the GS by applying
the view transform.
Then compute the face normal and the eye direction normal, both
in eye space.
Use the dot product of these two normals to test for back facing.
If not rejected, transform points to clip space with the projection
matrix and draw.